added image sharp for reducing image size

+++ utils/GCS.js
@@ -1,17 +1,38 @@
 import { bucket } from "../config/GCS.config.js";
+import sharp from "sharp";
 
-export const uploadToGCS = (file) => {
+export const uploadToGCS = async (file) => {
+  if (!["image/jpeg", "image/jpg", "image/png"].includes(file.mimetype)) {
+    throw new Error("Only JPG, JPEG, and PNG images are allowed");
+  }
+
+  let compressedBuffer;
+
+  if (file.mimetype === "image/jpeg" || file.mimetype === "image/jpg") {
+    compressedBuffer = await sharp(file.buffer)
+      .jpeg({ quality: 70, mozjpeg: true })
+      .toBuffer();
+  } else if (file.mimetype === "image/png") {
+    compressedBuffer = await sharp(file.buffer)
+      .png({ compressionLevel: 9 })
+      .toBuffer();
+  }
+
+  const ext = file.mimetype.split("/")[1];
+
+  const filename = `${Date.now()}-${file.originalname
+    .split(".")[0]
+    .replace(/\s+/g, "-")
+    .replace(/[^a-zA-Z0-9.-]/g, "")}.${ext}`;
+
+  const blob = bucket.file(filename);
+
+  const stream = blob.createWriteStream({
+    resumable: false,
+    contentType: file.mimetype,
+  });
+
   return new Promise((resolve, reject) => {
-    const filename = `${Date.now()}-${file.originalname
-      .replace(/\s+/g, "-")
-      .replace(/[^a-zA-Z0-9.-]/g, "")}`;
-    const blob = bucket.file(filename);
-
-    const stream = blob.createWriteStream({
-      resumable: false,
-      contentType: file.mimetype,
-    });
-
     stream.on("error", reject);
 
     stream.on("finish", () => {
@@ -23,11 +44,12 @@ export const uploadToGCS = (file) => {
       });
     });
 
-    stream.end(file.buffer);
+    stream.end(compressedBuffer);
   });
 };
 
 export const deleteFromGCS = async (fileName) => {
+  if (!fileName) throw new Error("File name is required");
   await bucket.file(fileName).delete();
 };
 
@@ -35,7 +57,7 @@ export const getSignedImageUrl = async (filename) => {
   const [url] = await bucket.file(filename).getSignedUrl({
     version: "v4",
     action: "read",
-    expires: Date.now() + 7 * 24 * 60 * 60 * 1000, // 7 days
+    expires: Date.now() + 7 * 24 * 60 * 60 * 1000,
   });
 
   return url;
